server/requirements.txt
fastapi>=0.104.1
uvicorn>=0.23.2
pydantic>=2.7.3
openai>=1.33.0
anthropic>=0.5.0
loguru>=0.7.0
python-dotenv>=1.0.0
pytest>=7.3.1
pytest-asyncio>=0.23.0
peewee>=3.16.0
psycopg2-binary>=2.9.6
aiohttp>=3.8.4
beautifulsoup4>=4.12.2
asyncio>=3.4.3
mem0ai>=0.1.81